adds WIP ttt frontend
